+++ src/inu/hardware/switch.py
@@ -14,17 +14,26 @@ class PullMode:
 
 
 class Switch:
-    def __init__(self, pin: int, pull: int = PullMode.PULL_DOWN, mode: str = SwitchMode.NO, on_change: callable = None):
+    def __init__(self, pin: int, pull: int = PullMode.PULL_DOWN, mode: str = SwitchMode.NO, min_active: int = 0,
+                 on_change: callable = None):
         """
         If the `mode` is NC, then an active current on the pin will be considered "off".
+        If `min_active` is set to non-zero, then the hardware must be in a new state for `min_active` ms before
+        changing state.
         """
         self.pin = Pin(pin, Pin.IN, pull=pull)
         self.state = None
         self.on_change = on_change
         self.reversed = mode == SwitchMode.NC
+        self.min_active = min_active
+
+        # Time the hardware changed state, we'll time until `min_active` ms then trigger a change
+        self.begin_state_change = None
+
+        # Time the device has been in an active state
         self.active_time = None
 
-    async def check_state(self) -> bool:
+    async def check_state(self, no_delay=False) -> bool:
         """
         Checks the state of the switch input pin. If the state has changed, `on_change(state)` will be called.
 
@@ -37,6 +46,16 @@ class Switch:
             state = not state
 
         if state != self.state:
+            if not no_delay and self.min_active > 0:
+                # We require a delay before changing state
+                if self.begin_state_change is None:
+                    self.begin_state_change = time.time()
+                    return self.state
+                else:
+                    if time.time() - self.begin_state_change < (self.min_active / 1000):
+                        return self.state
+
+            # State has changed, start an active time and trigger callbacks
             self.state = state
 
             if state:
@@ -46,6 +65,13 @@ class Switch:
 
             if self.on_change:
                 await self.on_change(self.state)
+
+        else:
+            # State is the same, check if we should reset the state-change counter
+            if self.begin_state_change is not None:
+                # This will happen if the device physically changes state quickly in succession before the min active
+                # time has expired (eg filter interference)
+                self.begin_state_change = None
 
         return self.state